backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
APScheduler==3.10.4
ebaysdk==2.2.0
python-dotenv==1.0.0
psycopg[binary]==3.2.3
alembic==1.13.1
gunicorn==21.2.0
